docs(goals): restate the ES5+untagged standalone goal as ~95.4% ex-dynamic-code
Project-lead ruling 2026-08-01 (option b, "for now" — revisitable).

The goal "100% ES5+untagged standalone" is arithmetically unreachable. Restated
as ~95.4%: target 8,150 of 8,545 reachable (6,176 passing + 1,974 non-dynamic
failures).

317 files are DECLINE-BY-DEPENDENCY and OUT OF SCOPE, not failures to fix:
  ~144 eval/Function  -> #2928 + packaging #2527 (real eval costs minutes to
       compile and is unaffordable per shard — a packaging problem as much as
       a semantics one)
  ~162 `with`         -> object environment records with first-class Reference
       identity; a front-end substrate on the same footing as the 795-file
       descriptor MOP
Near-disjoint; 13 need both. Funding eval does NOT deliver `with` — an earlier
version of the census said otherwise and has been corrected on main.

Consequences recorded so the next session inherits them: do not dispatch at the
317, do not measure against 8,545 or report "100%", and treat 95.4% as an UPPER
BOUND rather than a forecast (202 files remain unpriced).

The exclusion rests on a non-circularity control, not an assumption: the same
detector over the 6,176 goal-scope PASSES finds 248 files that use
eval/`with`/Function and pass anyway. The 317 were identified by engine refusal.

`with` is additionally 168 of 175 host-lane, so it is shared front-end
scope-analysis work rather than a standalone-gap item.

✓
